requirements.txt
discord-py-interactions==5.11.0
gspread==6.0.2
google-auth==2.29.0
python-dotenv==1.0.1
